+++ CHANGELOG.md
@@ -1,3 +1,5 @@
 # v2.14.24
 
 * `measure.time_remaining` will not return negative values
+* `MemoryErrorExceptionHandler` will use less memory procesing `MemoryError`s
+